+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Restored the `HubspotPrivateApp` connector, which was deleted by error in v6.0.0
+
 ## [6.0.0] 2024-03-12
 
 ### Changed